Refactor scroll position methods and variables in gallery
The scrollBehavior function in the router and the methods for saving and restoring last scroll position in the GalleryListPage have been commented out. Additionally, variable names related to scroll position storage in the GalleryListStore have been renamed to reflect the term 'galleryList' more accurately. This is aimed to streamline the code and make it easier to understand its purpose.