feat: Add Day.js UTC extension and weekly media upload count
The commit introduces Day.js with a UTC plugin to efficiently manage dates and times. This library is used to count how many media uploads have been made in the past 7 days, enhancing the media tracking functionality. The number of newly uploaded media within the week is displayed in MasterPage.vue for user reference.