feat: Update style and media support in GalleryDetailPage
This update expands media support on the GalleryDetailPage to include not only photos but videos as well. The style of the page has also been updated to display videos appropriately. The changes also include small modifications to the use of quotation marks and the removal of an unused style class.