feat: Update transaction types to boolean and update transaction UI form
This commit refactors the transaction types from an enum to a boolean in the `TransactionCreationData` interface. Corresponding changes have been made in the `TransactionCreateStore` and `TransactionListPage.vue`. This includes removing the `newTransactionData` from the page component and incorporating it directly in the store, as well as updating the transaction type input field in the form to radio buttons for 'Cash In' and 'Cash Out'.