feat: Initialize translation on button click
A new method, named initTranslation, has been added to the code. This method, when called, logs a message to the console. It has been attached to a click event on a button in the TranslationPage, enabling the translation initialization instruction to fire whenever this button is clicked. Additionally, the values in the LanguageCode enum have been updated to all uppercase.