feat: Implement debounce function for translation input
The 'TranslationPage.vue' has been updated to include a debounce function for the translation input. This function delays the invocation of the 'initTranslation' method which is called upon input event in the text field. This mechanism helps to limit the number of consecutive and potentially unnecessary calls to this method. Furthermore, importing and using 'onIonViewWillLeave' from 'Ionic/vue' ensures the debounce function is cleared when the user leaves the page.