feat: Add loading state to TranslationPage
A loading state is added to the TranslationPage component by using a new reactive variable, `isTranslating`. This variable is switched on at the start of the translation request and switched off upon its completion. It is also used to display a loading state in the textarea that shows the translated content.