ref: Simplify GalleryUploadStore and move reset operation to services
GalleryUploadStore has been simplified by removing complex operations and focusing only on maintaining state. The removed operations have now become responsibility of UploadMediaService. It ensures cleaner code organization and separation of concerns. Now, the reset operation is performed in the logout process on services, not on the store layer. This fits better with the architecture, given that services, not stores, are responsible for managing operations.